# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.2"
  title: SitOPT API Reference
# basePath prefixes all resource paths 
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - text/html

# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/html
paths:
...
  /situations/occured/{thing}/{situation}:
    x-swagger-router-controller: situations
    put:
      tags:
        - situation
      summary: Change "occured"-attribute
      description: Change the "occured"-attribute of either specified or all situations. For testing purpose only!
      operationId: situationOccured
      
      parameters:
        - in: path
          name: situation
          description: Name of situation.
          required: true
          type: string
        - in: path
          name: thing
          description: Name of the thing.
          required: true
          type: string
       

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetModel"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /situations/{thing}:
    x-swagger-router-controller: situations
    get:
      tags:
        - situation
      summary: Get all situations for a thing
      description: Get all situations for a specified thing.
      operationId: situationsByThing
      parameters:
        - name: thing
          in: path
          description: name of the thing
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/SituationModel"
        "404":
          description: Thing not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /situations/{thing}/{situation}:
    x-swagger-router-controller: situations
    get:
      tags:
        - situation
      summary: Get situation
      description: Get specified situations
      operationId: situationByThingAndTemplate
      
      parameters:
        - name: situation
          in: path
          description: name of situationtemplate
          required: true
          type: string
        - name: thing
          in: path
          description: name of the thing
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetModel"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
            
    delete:
      tags:
        - situation
      summary: Delete situation
      description: Deletes specified situation 
      operationId: deleteSituation
      
      parameters:
        - name: situation
          in: path
          description: name of the situation template
          required: true
          type: string
        - name: thing
          in: path
          description: name of the thing
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GetModel"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /situations/changes:
    x-swagger-router-controller: situations
    
    
    get:
      tags:
        - situation
      summary: Get registrations
      description: Specify CallbackURL to get only those registrations with the specified CallbackURL. If emtpy, returns all registrations.
      operationId: allRegistrations
      
      parameters:
        - name: CallbackURL
          in: query
          description: URL the changes are sent to.
          required: false
          type: string
          
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/RegistrationModel"
        # responses may fall through to errors
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"


    
    
    
    post:
      tags:
        - situation
      summary: Registrate for changes
      description: Get a notification when the situation - specified by corresponding thing & situationtemplate - changes via a callback to a specified URL. The parameter "continuous" controls whether the callback only happens once when the situation changes or continues forever. Either enter both IDs or none. 
      operationId: situationChange
      
      parameters:
        - name: SitTempName
          in: query
          description: name of situation template. If empty, any situation change will be sent to the callback URL
          required: false
          type: string
          
        - name: ThingName
          in: query
          description: name of thing. If empty, any situation change will be sent to the callback URL
          required: false
          type: string
          
        - name: CallbackURL
          in: query
          description: URL the callback is sent to
          required: true
          type: string
          
        - name: once
          in: query
          description: whether the callback continues or only once
          required: true
          type: boolean
          default: false

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "400":
          description: Bad Content
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    
    delete:        
      tags:
        - situation
      summary: Deletes registration
      description: Deletes the registration for either specified or all situation changes.
      
      operationId: situationChangeDelete
      
      parameters:
        - name: ID
          in: query
          description: ID of situation. If empty, callbacks to the specified URL will be stopped.
          required: false
          type: string
          
        - name: CallbackURL
          in: query
          description: URL to which the callback shall be stopped
          required: true
          type: string


      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /situations:
  # binds a127 app logic to a route
    x-swagger-router-controller: situations
    get:
      tags:
        - situation
      summary: Get all situations
      description: Get all situations
      operationId: allSituations
      
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/SituationModel"
        # responses may fall through to errors
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
      
    post:
      tags:
        - situation
      summary: Store situation
      description: Stores situation in database. ID of thing, situation template and sensors referenced in the situation must be in the database.
      operationId: saveSituation
      # used as the method name of the controller
      parameters:
        - in: body
          name: body
          description: input
          required: false
          schema:
            $ref: "#/definitions/SituationModel"
            
     
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /things:
  # binds a127 app logic to a route
    x-swagger-router-controller: things
    
    get:
      tags:
        - thing
      summary: Get all things
      description: Get all things
      operationId: allThings
      
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/ThingModel"
        # responses may fall through to errors
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
            
    post:
      tags:
        - thing
      summary: Save thing
      description: Save things in the database. Field "location" uses GeoJSON. Only use value of GeoJSON field "geometry"
      operationId: saveThing
      # used as the method name of the controller
      parameters:
        - in: body
          name: body
          description: input
          required: false
          schema:
            $ref: "#/definitions/ThingModel"
            
     
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  SituationModel:
    required:
      - thing
      - situationtemplate
      - occured
      - sensorvalues
    properties:
      thing:
        type: string
      timestamp:
        type: string
        default: ""
      situationtemplate:
        type: string
      occured:
        type: boolean
      sensorvalues:
        type: array
        items:
          $ref: "#/definitions/SensorvalueModel"
  GetModel:
    required:
      - name
    properties:
      name:
        type: string
  ThingModel:
    required:
      - name
      - sensors
      - owners
    properties:
      state:
        type: array
        default: []
        items:
          type: string
      name:
        type: string
      url:
        type: string
        default: ""
      description:
        type: string
        default: ""
      sensors:
        type: array
        items:
          type: string
      location:
        type: string
        default: ""
      owners:
        type: array
        items:
          type: string
...
  SensorvalueModel:
    required:
      - sensor
      - value
    properties:
      sensor:
        type: string 
      value:
        type: number
        format: float
      timestamp:
        type: string
        default: ""
      quality:
        type: number
        format: float
        default: 0
  RegistrationModel:
    required:
      - ThingID
      - TemplateID
      - callbackURL
      - once
    properties:
      ThingID:
        type: string
      TemplateID:
        type: string
      callbackURL:
        type: string
      once:
        type: boolean
  Response:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string